Create custom hook for logic from CurrentGameplay
